+++ HotelWebsite/Startup.cs
@@ -7,6 +7,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Service;
+using Service.Reviews;
 using System;
 using System.Threading.Tasks;
